requirements.txt
# =============================================================================
# Dependencies - Flask Deploy Template (Production)
# =============================================================================
# Run: pip install -r requirements.txt
# Audit: pip-audit
# =============================================================================

# Core Framework
Flask==3.0.0
Werkzeug==3.1.5
Jinja2==3.1.6

# Security & Forms
flask-wtf==1.2.1

# Rate Limiting (optional, for local development)
flask-limiter==3.5.0

# Configuration
python-dotenv==1.0.0
